cannot pay if nothing in cart

+++ client/src/Components/Cart.js
@@ -1,6 +1,10 @@
 import React, {useState, useEffect} from 'react'
+import StripeCheckout from 'react-stripe-checkout';
 import {ListGroup, ListGroupItem, Row, Col, Button} from 'reactstrap';
 import axios from 'axios';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
+import { faArrowRight } from '@fortawesome/free-solid-svg-icons';
+import {Link} from 'react-router-dom';
 
 const Cart = () => {
   const [products, setProducts] = useState([])
@@ -11,6 +15,29 @@ const Cart = () => {
       setProducts(res.data)
     })
   }, []);
+
+  // remove all items from cart after purchase
+  const clearCart = () => {
+    axios.delete(`api/items/`).then(response => {
+      const cartList = []
+      setProducts(cartList);
+    });
+  }
+
+  // handles stripe token
+  const handleToken = async (token) => {
+    const response = await axios.post('/api/items/checkout', {
+      token
+    });
+    const {status} = response.data
+    if(status === 'success') {
+      console.log('this was a successful purchase')
+      clearCart()
+      window.location.href = response.data.redirect
+    } else {
+      console.log('poooop. this did not work')
+    }
+  }
 
   // delete item from database function
   const deleteProduct = (productId) => {
@@ -29,34 +56,58 @@ const Cart = () => {
     return total
   }
  
-  return(
-    <div>
-      <ListGroup className="pb-5">
-        {products.map(({name, price, image, _id}) => (
-          <ListGroupItem key={_id}>
-            <Row>
-              <Col className="col-12 col-md-2">
-                <img style={{width: '100%'}} src={image} alt="the product"/>
-              </Col>
-              <Col className="col-12 col-md-8">
-                <p className="text-muted"><strong>{name}</strong></p>
-                <Button
-                  className="bg-danger"
-                  onClick={() => deleteProduct(_id)}
-                >
-                  delete
-                </Button>
-              </Col>
-              <Col style={{color: 'black'}}  className="col-12 col-md-2"><strong>${price}</strong></Col>
-            </Row>
-          </ListGroupItem>
-        ))}
-      </ListGroup>
+  if(products.length === 0) {
+    return(
+      <div>
+        <h1>Cart is empty</h1>
+        <Link to="/" className="btn btn-primary"><FontAwesomeIcon
+        icon={faArrowRight}/> Back Home</Link>
+      </div>
+    )
+  } else {
+    return(
+      <Row>
+        <Col className="col-12 col-md-10">
+        <ListGroup className="pb-5">
+          {products.map(({name, price, image, _id}) => (
+            <ListGroupItem className="bg-dark text-white" key={_id}>
+              <Row>
+                <Col className="col-12 col-md-2">
+                  <img style={{width: '100%'}} src={image} alt="the product"/>
+                </Col>
+                <Col className="col-12 col-md-8">
+                  <p className="text-white"><strong>{name}</strong></p>
+                  <Button
+                    className="bg-danger"
+                    onClick={() => deleteProduct(_id)}
+                  >
+                    delete
+                  </Button>
+                </Col>
+                <Col style={{color: 'white'}}  className="col-12 col-md-2"><strong>${price}</strong></Col>
+              </Row>
+            </ListGroupItem>
+          ))}
+        </ListGroup>
+        <hr className="bg-light"/>
         <p className="text-right">
           Total ({products.length} items): ${totalPrice()}
         </p>
-    </div>
-  )
+        </Col>
+        <Col className="col-12 col-md-2">
+          <StripeCheckout
+            stripeKey="pk_test_SDgPMwlzPSK2Fo1LNTcCKhYr00pduFRDkx"
+            token={handleToken}
+            billingAddress
+            shippingAddress
+            amount={2500}
+            name="the dishwasher"
+          />
+        </Col>
+      </Row>
+    )
+  }
+  
 }
 
 export default Cart;
